+++ SpringBootPractice-User_DM/src/main/java/com/example/demo/user/config/DatabaseConfig.java
@@ -1,0 +1,38 @@
+package com.example.demo.user.config;
+
+import javax.sql.DataSource;
+
+import org.apache.ibatis.session.SqlSessionFactory;
+import org.mybatis.spring.SqlSessionFactoryBean;
+import org.mybatis.spring.SqlSessionTemplate;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.core.io.support.PathMatchingResourcePatternResolver;
+
+import com.example.demo.user.model.UserVO;
+
+
+@Configuration
+public class DatabaseConfig {
+
+	@Bean
+    public SqlSessionFactory sqlSessionFactory(DataSource dataSource) throws Exception {
+        final SqlSessionFactoryBean sessionFactory = new SqlSessionFactoryBean();
+        sessionFactory.setDataSource(dataSource);
+        sessionFactory.setTypeAliasesPackage("com.example.demo.user.model");
+        Class[] aliases = {UserVO.class};
+        sessionFactory.setTypeAliases(aliases);
+        PathMatchingResourcePatternResolver resolver = new PathMatchingResourcePatternResolver();
+        sessionFactory.setMapperLocations(resolver.getResources("classpath:mapper/**.xml"));
+        return sessionFactory.getObject();
+    }
+
+	@Bean
+    public SqlSessionTemplate sqlSessionTemplate(SqlSessionFactory sqlSessionFactory) throws Exception {
+      final SqlSessionTemplate sqlSessionTemplate = new SqlSessionTemplate(sqlSessionFactory);
+      return sqlSessionTemplate;
+    }
+
+
+
+}
