+++ SpringBootPractice-User_DM/src/main/resources/application.properties
@@ -1,4 +1,4 @@
-server.port=8080
+server.port=8989
 
 spring.mvc.view.prefix=/WEB-INF/view/
 spring.mvc.view.suffix=.jsp
