+++ SpringBootPractice-User_DM/src/main/resources/mapper/UserMapper.xml
@@ -3,7 +3,7 @@
 PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
 "http://mybatis.org/dtd/mybatis-3-mapper.dtd" >
 
-<mapper namespace="com.example.demo.user.mapper.UserMapper">
+<mapper namespace="mapper.UserMapper">
 	
 	<select id="userLogin" parameterType="String"
 		resultType="com.example.demo.user.model.UserVO">
